Controller 공통 로직 추상 클래스로 추출 AbstractFrontController 도입
