+++ src/main/java/com/hta2405/unite/controller/AdminFrontController.java
@@ -1,85 +1,23 @@
 package com.hta2405.unite.controller;
 
-import com.hta2405.unite.action.*;
 import com.hta2405.unite.action.admin.*;
-import jakarta.servlet.RequestDispatcher;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
-import jakarta.servlet.http.HttpServlet;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-import java.io.IOException;
-import java.io.Serial;
-import java.util.HashMap;
 
 @WebServlet("/admin/*")
-public class AdminFrontController extends HttpServlet {
-    @Serial
-    private static final long serialVersionUID = 1L;
-    HashMap<String, Action> actionMap = new HashMap<>();
+public class AdminFrontController extends AbstractFrontController {
 
-    //아래에 URL, Action 추가
     @Override
     public void init() throws ServletException {
-        actionMap.put("/holiday", new AdminHolidayAction()); //휴일 관리
-        actionMap.put("/holiday/insert", new AdminHolidayInsertAction()); //휴일 직접 등록
-        actionMap.put("/holiday/delete", new AdminHolidayDeleteAction()); //휴일 직접 등록
-        actionMap.put("/holiday/api", new AdminHolidayApiAction()); //공휴일 api로 받아와서 DB 업데이트
-        actionMap.put("/holiday/weekend", new AdminHolidayWeekendAction()); //주말 DB에 업데이트
-        actionMap.put("/popup", new AdminPopupAction()); //팝업 등록
-        actionMap.put("/resource", new AdminResourceAction()); //자원관리
-        actionMap.put("/resource/add", new AdminResourceAddAction()); //자원관리-추가
-        actionMap.put("/resource/edit", new AdminResourceEditAction()); //자원관리-수정
-        actionMap.put("/resource/delete", new AdminResourceDeleteAction()); //자원관리-삭제
-    }
-
-    protected void doProcess(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        System.out.println("요청 주소 : " + req.getRequestURL());
-
-        String requestURI = req.getRequestURI();
-        System.out.println("requestURI = " + requestURI);
-
-        String contextPath = req.getContextPath();
-        System.out.println("contextPath = " + contextPath);
-
-        String command = requestURI.substring(contextPath.length() + "/admin".length());
-        System.out.println("command = " + command);
-
-        //등록된 URL이 아닌경우 404에러페이지 보여줌
-        if (!actionMap.containsKey(command)) {
-            RequestDispatcher dispatcher = req.getRequestDispatcher("/error/404.jsp");
-            dispatcher.forward(req, resp);
-            return;
-        }
-
-        //초기화
-        Action action = actionMap.get(command);
-
-        //요청 주소 별 ActionForward(리다이렉트 여부, 다음 주소 정보)객체 생성해서 변수에 담음
-
-        //해시에 저장되어있는 new 객체들은 생성은 되지만, 이때 비로소 'forward' 로 참조변수가 설정됨
-        //action.execute 로 리다이렉트 여부, 주소를 forward 변수에 담고, request 변수에 model 객체를 담음
-        ActionForward forward = action.execute(req, resp);
-
-        if (forward != null) {
-            if (forward.isRedirect()) {
-                resp.sendRedirect(forward.getPath());
-            } else {
-                //null을 반환할 경우
-                RequestDispatcher dispatcher = req.getRequestDispatcher(forward.getPath());
-                dispatcher.forward(req, resp);
-            }
-        }
-    }
-
-    @Override
-    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        doProcess(req, resp);
-    }
-
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        doProcess(req, resp);
+        actionMap.put("/holiday", new AdminHolidayAction());
+        actionMap.put("/holiday/insert", new AdminHolidayInsertAction());
+        actionMap.put("/holiday/delete", new AdminHolidayDeleteAction());
+        actionMap.put("/holiday/api", new AdminHolidayApiAction());
+        actionMap.put("/holiday/weekend", new AdminHolidayWeekendAction());
+        actionMap.put("/popup", new AdminPopupAction());
+        actionMap.put("/resource", new AdminResourceAction());
+        actionMap.put("/resource/add", new AdminResourceAddAction());
+        actionMap.put("/resource/edit", new AdminResourceEditAction());
+        actionMap.put("/resource/delete", new AdminResourceDeleteAction());
     }
 }
